+++ examples/modules/key-pair/outputs.tf
@@ -1,0 +1,4 @@
+output "key_pair" {
+  description = "Generated key pair."
+  value       = aws_key_pair.key
+}