+++ examples/runner-public/main.tf
@@ -24,13 +24,22 @@ module "cache" {
   environment = "${var.environment}"
 }
 
+
+
+module "key_pair" {
+  source = "../modules/key-pair"
+
+  environment = var.environment
+  name        = var.runner_name
+}
+
 module "runner" {
   source = "../../"
 
   aws_region  = var.aws_region
   environment = var.environment
 
-  ssh_public_key              = local_file.public_ssh_key.content
+  ssh_key_pair                = module.key_pair.key_pair.key_name
   runners_use_private_address = false
 
   vpc_id                   = module.vpc.vpc_id
@@ -77,7 +86,7 @@ module "runner2" {
   aws_region  = var.aws_region
   environment = "${var.environment}-2"
 
-  ssh_public_key              = local_file.public_ssh_key.content
+  ssh_key_pair                = module.key_pair.key_pair.key_name
   runners_use_private_address = false
 
   vpc_id                   = module.vpc.vpc_id